pyproject.toml
[project]
name = "kiteclient"
version = "1.0.0"
description = "UNofficial Kite Connect API Python client"
readme = "README.md"
authors = [{ name = "Navdeep Virk", email = "nsvirk@gmail.com" }]
requires-python = ">=3.13"
dependencies = ["pyotp>=2.9.0", "requests>=2.32.4"]

keywords = ["kite", "api", "client", "trading", "finance"]

[project.scripts]
kiteclient = "kiteclient:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["kiteclient"]

[dependency-groups]
dev = [
    "python-dotenv>=1.1.1",
]
